Move interfaces to upper level.
