+++ src/ispd/motor/carga/impl/CollectionWorkloadGenerator.java
@@ -1,5 +1,7 @@
 package ispd.motor.carga.impl;
 
+import ispd.motor.carga.WorkloadGenerator;
+import ispd.motor.carga.WorkloadGeneratorType;
 import ispd.motor.filas.RedeDeFilas;
 import ispd.motor.filas.Tarefa;
