+++ src/ispd/motor/carga/impl/GlobalWorkloadGenerator.java
@@ -1,5 +1,6 @@
 package ispd.motor.carga.impl;
 
+import ispd.motor.carga.WorkloadGeneratorType;
 import ispd.motor.filas.RedeDeFilas;
 import ispd.motor.filas.Tarefa;
 import ispd.motor.filas.servidores.CS_Processamento;
