+++ src/ispd/motor/carga/impl/RandomicWorkloadGenerator.java
@@ -1,5 +1,6 @@
 package ispd.motor.carga.impl;
 
+import ispd.motor.carga.WorkloadGenerator;
 import ispd.motor.carga.task.TaskBuilder;
 import ispd.motor.random.TwoStageUniform;
 import ispd.motor.random.Distribution;
