+++ src/ispd/motor/carga/task/TaskBuilder.java
@@ -1,6 +1,6 @@
 package ispd.motor.carga.task;
 
-import ispd.motor.carga.impl.WorkloadGenerator;
+import ispd.motor.carga.WorkloadGenerator;
 import ispd.motor.filas.RedeDeFilas;
 import ispd.motor.filas.Tarefa;
 import ispd.motor.filas.servidores.CS_Processamento;
